+++ cognite/experimental/_api/entity_extraction.py
@@ -1,13 +1,12 @@
 from typing import List
 
 from cognite.experimental._context_client import ContextAPI
-from cognite.experimental.data_classes import ContextualizationJob
 
 
 class EntityExtractionAPI(ContextAPI):
     _RESOURCE_PATH = "/context/entity_extraction"
 
-    def extract(self, file_ids: List[int], entities: List[str]) -> "Task[ContextualizationJob]":
+    def extract(self, file_ids: List[int], entities: List[str]) -> "ContextualizationJob":
         """Extracts entities from files.
 
         Args:
@@ -15,5 +14,5 @@ class EntityExtractionAPI(ContextAPI):
             entities: Entities to search for.
 
         Returns:
-            Task[ContextualizationJob]: Task which waits for the job to be completed."""
-        return self._run_job(job_path="/extract", file_ids=file_ids, entities=entities)
+            ContextualizationJob: Resulting queued job. Note that .results property of this job will block waiting for results."""
+        return self._run_job(job_path="/extract", status_path="/", file_ids=file_ids, entities=entities)