+++ cognite/experimental/_api/entity_matching.py
@@ -5,25 +5,25 @@ from cognite.experimental.data_classes import ContextualizationJob, EntityMatchi
 
 
 class EntityMatchingAPI(ContextModelAPI):
-    _RESOURCE_PATH = "/context/entity_matching"
+    _RESOURCE_PATH = EntityMatchingModel._RESOURCE_PATH
     _MODEL_CLASS = EntityMatchingModel
 
-    def fit(self, items: List[str]) -> "Task[EntityMatchingModel]":
+    def fit(self, items: List[str]) -> EntityMatchingModel:
         """Fit entity matching model.
 
         Args:
             items: list of entities to create a model for.
 
         Returns:
-            Task[EntityMatchingModel]: Task which waits for the model to be completed."""
+            EntityMatchingModel: Resulting queued model."""
         return super()._fit_model(items=list(items))
 
-    def create_rules(self, matches: List[Dict]) -> "Task[ContextualizationJob]":
+    def create_rules(self, matches: List[Dict]) -> ContextualizationJob:
         """Fit rules model.
 
         Args:
             matches: list of matches to create rules for, given as a dictionary of from: to matches.
 
         Returns:
-            Task[ContextualizationJob]: Task which waits for the job to be completed."""
-        return self._run_job(job_path="/rules", status_path="/rules/", items=matches)
+            ContextualizationJob: Resulting queued job. Note that .results property of this job will block waiting for results."""
+        return self._run_job(job_path="/rules", items=matches)