+++ tests/tests_integration/test_contextualization/test_resource_typing.py
@@ -1,5 +1,3 @@
-import asyncio
-
 import pytest
 
 from cognite.experimental import CogniteClient
@@ -10,21 +8,16 @@ RTAPI = COGNITE_CLIENT.resource_typing
 
 
 class TestResourceTypingIntegration:
-    @pytest.mark.skip("hangs for some reason")
-    @pytest.mark.asyncio
-    async def test_fit(self):
+    def test_fit(self):
         items = [{"data": ["a", "b'"], "target": "x"}, {"data": ["c", "d'"], "target": "y"}] * 2
         targets_to_classify = ["x"]
 
-        resp = RTAPI.fit(items, targets_to_classify=targets_to_classify)
-        assert isinstance(resp, asyncio.Task)
-        model = await resp
+        model = RTAPI.fit(items, targets_to_classify=targets_to_classify)
         assert isinstance(model, ResourceTypingModel)
+        assert "Queued" == model.status
+        job = model.predict(items=[{"data": ["a", "b'"]}])
         assert "Completed" == model.status
-        jt = model.predict(items=[{"data": ["a", "b'"]}])
-        assert isinstance(jt, asyncio.Task)
-        job = await jt
         assert isinstance(job, ContextualizationJob)
-        assert isinstance(job.items, list)
-        assert {"data", "score", "target"} == job.items[0].keys()
+        assert isinstance(job.result["items"], list)
+        assert {"data", "score", "target"} == job.result["items"][0].keys()
         RTAPI.delete(model)